refactor: move AIChat state management to a zustand store
